feat: implement staff management service layer, DTOs, and base department and designation services

+++ src/main/java/com/schoolerp/staff/dto/StaffCreateRequest.java
@@ -59,7 +59,7 @@ public record StaffCreateRequest(
         String country,
 
         String postalCode,
-        Integer roleId,
+        Long roleId,
 
         /* ===== QUALIFICATIONS ===== */
